Enable linters on test files; fix errors (#549)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v           ✰  Thanks for creating a PR! You're awesome! ✰
v Please note that maintainers will only review those PRs with a
completed PR template.
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Purpose of Changes and their Description

Enable linters on test files; fix errors

## Link(s) to Ticket(s) or Issue(s) resolved by this PR


https://linear.app/alloralabs/issue/PROTO-2247/include-test-files-on-linter-checks

## Are these changes tested and documented?

Yes.

## Still Left Todo

*Fill this out if this is a Draft PR so others can help.*

+++ x/emissions/keeper/queryserver/query_server_losses_test.go
@@ -27,7 +27,7 @@ func (s *KeeperTestSuite) TestGetNetworkLossBundleAtBlock() {
 		ctx,
 		&types.QueryNetworkLossBundleAtBlockRequest{
 			TopicId:     topicId,
-			BlockHeight: int64(blockHeight),
+			BlockHeight: blockHeight,
 		},
 	)
 
@@ -102,7 +102,7 @@ func (s *KeeperTestSuite) TestGetReputerLossBundlesAtBlock() {
 
 	req := &types.QueryReputerLossBundlesAtBlockRequest{
 		TopicId:     topicId,
-		BlockHeight: int64(block),
+		BlockHeight: block,
 	}
 	response, err := s.queryServer.GetReputerLossBundlesAtBlock(ctx, req)
 	require.NoError(err)
